src-python/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-socketio>=5.11.0
sqlalchemy>=2.0.0
watchdog>=4.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
